Refactor SearchPage route config. Export loadData from SearchPage.duck.js file
