+++ src/containers/SearchPage/SearchPage.js
@@ -17,7 +17,7 @@ import { manageDisableScrolling, isScrollingDisabled } from '../../ducks/UI.duck
 import { SearchMap, ModalInMobile, Page } from '../../components';
 import { TopbarContainer } from '../../containers';
 
-import { searchListings, searchMapListings, setActiveListing } from './SearchPage.duck';
+import { searchMapListings, setActiveListing } from './SearchPage.duck';
 import {
   pickSearchParamsOnly,
   validURLParamsForExtendedData,
@@ -27,10 +27,6 @@ import {
 import MainPanel from './MainPanel';
 import css from './SearchPage.module.css';
 
-// Pagination page size might need to be dynamic on responsive page layouts
-// Current design has max 3 columns 12 is divisible by 2 and 3
-// So, there's enough cards to fill all columns on full pagination pages
-const RESULT_PAGE_SIZE = 24;
 const MODAL_BREAKPOINT = 768; // Search is in modal on mobile layout
 const SEARCH_WITH_MAP_DEBOUNCE = 300; // Little bit of debounce before search is initiated.
 
@@ -311,25 +307,5 @@ const SearchPage = compose(
   ),
   injectIntl
 )(SearchPageComponent);
-
-SearchPage.loadData = (params, search) => {
-  const queryParams = parse(search, {
-    latlng: ['origin'],
-    latlngBounds: ['bounds'],
-  });
-  const { page = 1, address, origin, ...rest } = queryParams;
-  const originMaybe = config.sortSearchByDistance && origin ? { origin } : {};
-  return searchListings({
-    ...rest,
-    ...originMaybe,
-    page,
-    perPage: RESULT_PAGE_SIZE,
-    include: ['author', 'images'],
-    'fields.listing': ['title', 'geolocation', 'price'],
-    'fields.user': ['profile.displayName', 'profile.abbreviatedName'],
-    'fields.image': ['variants.landscape-crop', 'variants.landscape-crop2x'],
-    'limit.images': 1,
-  });
-};
 
 export default SearchPage;
