+++ src/containers/pageDataLoadingAPI.js
@@ -1,9 +1,14 @@
 /**
  * Export loadData calls from ducks modules of different containers
  */
+import { loadData as SearchPageLoader } from './SearchPage/SearchPage.duck';
 
 const getPageDataLoadingAPI = () => {
-  return {};
+  return {
+    SearchPage: {
+      loadData: SearchPageLoader,
+    },
+  };
 };
 
 export default getPageDataLoadingAPI;
